+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.14] - 2023-11-4
+
+This adds support for set returning functions for unspecified record
+types, and it also changes the type used to represent a set of rows
+from a vector to a list.
+
 ## [0.5.13] - 2023-10-31
 
 This adds support for set returning functions, but only when it's a